+++ src/game/Game.h
@@ -4,6 +4,11 @@
 #include "game/GameWindow.h"
 
 #include <teya/graphics/PixelCanvas.h>
+#if TEYA_ENABLE_EDITOR
+#include "game/GameEditorHost.h"
+#include <teya/editor/Editor.h>
+#include <memory>
+#endif
 
 class Game {
   public:
@@ -16,6 +21,15 @@ class Game {
     Game &operator=(Game &&) = delete; // Move assignment
 
     void run();
+#if TEYA_ENABLE_EDITOR
+    RenderTexture2D editorTexture() const;
+    int editorCanvasWidth() const;
+    int editorCanvasHeight() const;
+    std::vector<teya::editor::RuntimeNode> editorHierarchy() const;
+    std::vector<teya::editor::RuntimeProperty> editorProperties(teya::editor::RuntimeObjectId id) const;
+    teya::editor::EditorFrameMetrics editorMetrics() const;
+    void requestEditorExit();
+#endif
 
   private:
     void update(float deltaTime);
@@ -25,4 +39,9 @@ class Game {
     teya::graphics::PixelCanvas canvas_;
     GameStateMachine gameStates_;
     bool exitRequested_ = false;
+#if TEYA_ENABLE_EDITOR
+    std::unique_ptr<GameEditorHost> editorHost_;
+    std::unique_ptr<teya::editor::Editor> editor_;
+    teya::editor::EditorFrameMetrics metrics_;
+#endif
 };